refactor: move general entity stuf to ngrx-data
