+++ src/client/app/store/actions/index.ts
@@ -1,3 +1,1 @@
-export * from './data.actions';
-export * from './entity.actions';
 export * from './hero.actions';
