+++ src/client/app/store/effects/hero.effects.ts
@@ -7,14 +7,15 @@ import { concatMap, catchError, first, map, mergeMap, switchMap } from 'rxjs/ope
 import { Action, Store } from '@ngrx/store';
 import { Actions, Effect } from '@ngrx/effects';
 
+import { EntityCache, toAction } from '../ngrx-data';
 import * as HeroActions from '../actions';
 
 import { Hero } from '../../model';
-import { HeroDataService, DataServiceError } from '../services';
-import { EntityCache } from '../reducers';
+import { HeroDataService } from '../services';
 
 const filterAction = new HeroActions.GetFilteredHeroes();
-const toAction = HeroActions.toAction(filterAction);
+const toHeroAction = toAction(filterAction);
+
 type HeroAction = HeroActions.HeroAction;
 
 @Injectable()
@@ -23,52 +24,44 @@ export class HeroEffects {
   getHeroes$: Observable<Action> = this.actions$
     .ofType(HeroActions.GET_HEROES)
     .pipe(
-      switchMap(() =>
-        toAction(
-          this.heroDataService.getHeroes(),
-          HeroActions.GetHeroesSuccess,
-          HeroActions.GetHeroesError
-        )
-      )
+      switchMap(() => toHeroAction(
+        this.heroDataService.getHeroes(),
+        HeroActions.GetHeroesSuccess,
+        HeroActions.GetHeroesError
+      ))
     );
 
   @Effect()
   addHero$: Observable<Action> = this.actions$
     .ofType(HeroActions.ADD_HERO)
     .pipe(
-      concatMap((action: HeroAction) =>
-        toAction(
-          this.heroDataService.addHero(action.payload),
-          HeroActions.AddHeroSuccess,
-          HeroActions.AddHeroError
-        )
-      )
+      concatMap((action: HeroAction) => toHeroAction(
+        this.heroDataService.addHero(action.payload),
+        HeroActions.AddHeroSuccess,
+        HeroActions.AddHeroError
+      ))
     );
 
   @Effect()
   deleteHero$: Observable<Action> = this.actions$
     .ofType(HeroActions.DELETE_HERO)
     .pipe(
-      concatMap((action: HeroAction) =>
-        toAction(
-          this.heroDataService.deleteHero(action.payload),
-          HeroActions.DeleteHeroSuccess,
-          HeroActions.DeleteHeroError
-        )
-      )
+      concatMap((action: HeroAction) => toHeroAction(
+        this.heroDataService.deleteHero(action.payload),
+        HeroActions.DeleteHeroSuccess,
+        HeroActions.DeleteHeroError
+      ))
     );
 
   @Effect()
   updateHero$: Observable<Action> = this.actions$
     .ofType<HeroActions.UpdateHero>(HeroActions.UPDATE_HERO)
     .pipe(
-      concatMap((action: HeroAction) =>
-        toAction(
-          this.heroDataService.updateHero(action.payload),
-          HeroActions.UpdateHeroSuccess,
-          HeroActions.UpdateHeroError
-        )
-      )
+      concatMap((action: HeroAction) => toHeroAction(
+        this.heroDataService.updateHero(action.payload),
+        HeroActions.UpdateHeroSuccess,
+        HeroActions.UpdateHeroError
+      ))
     );
 
   constructor(
