+++ src/client/app/store/ngrx-data/index.ts
@@ -1,0 +1,4 @@
+export * from './entity.actions';
+export * from './entity.effects';
+export * from './entity.reducer';
+export * from './interfaces';
