+++ src/client/app/store/services/data.service.ts
@@ -2,6 +2,3 @@
 export const api = '/api';
 export const fakeDelays = { select: 1000, save: 200 };
 
-export class DataServiceError<T> {
-  constructor(public error: any, public requestData: T) {}
-}
