+++ src/client/app/store/services/index.ts
@@ -1,4 +1,3 @@
-export { DataServiceError } from './data.service';
 export * from './hero-data.service';
 export * from './hero.dispatchers';
 export * from './hero.selectors';
